refactor: multer on users routes
